api/requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
supabase-py==0.0.6
pandas==1.3.3
openpyxl==3.0.7
smtplib
email